requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit
langchain
langchain-community
langchain-openai
torch
transformers
pandas
python-dotenv
